Move to v2.8. Fix tests issues.
