+++ src/main/java/org/cloudfoundry/community/servicebroker/postgresql/Application.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.cloudfoundry.community.servicebroker.postgresql.config;
+package org.cloudfoundry.community.servicebroker.postgresql;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.EnableAutoConfiguration;
