+++ src/test/resources/application.properties
@@ -1,4 +1,4 @@
 security.user.password: password
-service_id: pg
+service_id: d90aa98f-eead-478e-bde3-de692f66be45
 plan_id: postgresql-basic-plan
-MASTER_JDBC_URL: jdbc:postgresql://localhost:5432/travis_ci_test?user=postgres&password=
+MASTER_JDBC_URL: jdbc:postgresql://localhost:5432/cfsb?user=cfsb&password=
